Added to visual plug-in over Vertex icon drawing.

+++ payola/s2js/adapters/src/main/scala/s2js/adapters/js/dom/CanvasPixelArray.scala
@@ -5,8 +5,8 @@ abstract class CanvasPixelArray
     val length: Long
 
     // TODO: octet retval
-    def getter(index: Long): Long
+    def getter(index: Long): Double
 
     // TODO: octet value
-    def setter(index: Long, value: Long)
+    def setter(index: Long, value: Double)
 }
